tach timestamp va date ra dap ung audit thơi gian

+++ movie-service/src/main/java/com/example/movieservice/entity/MovieRole.java
@@ -32,6 +32,6 @@ public class MovieRole {
     private Long movie;
 
     @Column(name = "role")
-    private Integer role;
+    private Long role;
 
 }
